Add management :
- for CAP alignement not sure if it is technicaly good, but visualy it
is
- for begin and end figure

these are needed, because there is an issue with this lib : the filling
of polybezier is never done.

I noticed that Polybezier which are inside figure, are closed shape, and
so a Genaralpath must be used instead of a basic shape.


I added a error message warning the user when encoutering a tiff